Added imageindexer and searcher
Also added a simple flask bootstrap web app
which can upload an image file to be used by
our imagesearcher.

Searching is done by calculating the scores
of all frames of a video. Where every video
score is incremeted with 1 / (image_score + 1).
This way we can sort the videos
by most relevant to the picture.